Set up twitch-openvla-client connection. (#4)
Problem: even the franka arm is kind of bad. It is incapable of doing
simple tasks in sim most of the time, although I am unsure as to why.

I will try playing around with some parameters (e.g. unnorm_key) to see
if performance can be improved.

Additionally the process crashes if each run of the model is not
truncated at a sufficiently short length. It takes the computer out
along with it, which is far from ideal.